moving dummy encoding to Imputer (still need to update documentation)
