+++ microimpute/comparisons/__init__.py
@@ -8,7 +8,6 @@ from .autoimpute import autoimpute
 
 # Import data handling functions
 from .data import (
-    postprocess_imputations,
     prepare_scf_data,
     preprocess_data,
     scf_url,
